+++ packages/next-templates/src/app/wmebv/niet-ingelogd/stap4/stap4.tsx
@@ -1,0 +1,20 @@
+'use client';
+
+import { UtrechtArticle, UtrechtPage, UtrechtPageContent } from '@utrecht/web-component-library-react';
+import { ExampleHeaderWmebv } from '@/components/ExampleHeader/wmebv/ExampleHeaderWmebv';
+import { ExampleFooter } from '@/components/ExampleFooter/ExampleFooter';
+import { ExampleNavigation } from '@/components/ExampleNavigation/ExampleNavigation';
+import '@/app/styling/css/wmebv.css';
+
+export default function home() {
+  return (
+    <UtrechtPage>
+      <ExampleHeaderWmebv />
+      <ExampleNavigation />
+      <UtrechtPageContent>
+        <UtrechtArticle></UtrechtArticle>
+      </UtrechtPageContent>
+      <ExampleFooter />
+    </UtrechtPage>
+  );
+}
